requirements.txt
asgiref==3.4.1
Django==3.2.6
Markdown==3.3.4
python-frontmatter==1.0.0
pytz==2021.1
PyYAML==5.1
sqlparse==0.4.1
uwsgi
